Add option to only perform replacements. Useful for cleaning OCR output, as most mistakes are replacements.