@fsondej 2021-01-19
Code looks great, good job!
Please also add some short unit test in test_all.py and a short section in README, maybe  `# OCR` to document the change.