@Jerry2001Qu 2021-01-19
Awesome, I've added a section in the README, and a unit test.